Style Driver Profile modal
