+++ src/styles/components/DriverDetails.scss
@@ -1,0 +1,34 @@
+.info-driver-box {
+  border: 10px solid rgba(35, 37, 38, 1);
+  margin: 1rem;
+  padding: 0.5rem;
+  font-family: 'Orbitron', sans-serif;
+
+  .info-driver-div {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 1rem;
+    gap: 1rem;
+
+    .driver-big-picture {
+      height: 300px;
+      width: auto;
+    }
+
+    .driver-info-details {
+      height: 320px;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-around;
+
+      span {
+        font-size: 1rem;
+        padding: 1rem 0 1rem 1rem;
+        margin: 0.3rem;
+        width: 100%;
+        border: 1px solid black;
+      }
+    }
+  }
+}
